Changed from using a log and linear approximation for LL to two linears.
Tried removing all LL points (after corrected with equations, and
differences taken) where there was less than 2.5L already in the
bucket. Cutoff in MM for this determined by the single calibration we
did on each LL. Turns out the measurements from the field never
registered that little water. Actually, 2.5L is too small, only at 3 L
do we start to notice an appreciable slope that shows differences in
the water level.  We need to go up to 3 litres. The data in the field
doesn’t actually change at our definition (in MM) of 3L, so let’s use
the repeated LL calibrations to find an averaged height in MM that
better corresponds to 3L for the field data. It turns out that when we
want to define the lower correction equation for just 3-4 L, there
aren’t enough points to do this with a logarithmic equation, so it’s
necessary to change the lower equation to linear.